backend/requirements.txt
fastapi>=0.128.0
sqlmodel>=0.0.31
uvicorn>=0.40.0
psycopg2-binary>=2.9.11
python-dotenv>=1.0.0
better-auth>=0.0.1b11
PyJWT>=2.8.0
python-json-logger>=2.0.7
bcrypt>=4.0.1
openai>=1.0.0
